修改FieldSizeAttribute特性，允许字节作为存储长度的类型

+++ NewLife.Core/Serialization/Binary/FieldSizeAttribute.cs
@@ -71,7 +71,7 @@ namespace NewLife.Serialization
 
             // 目标字段必须是整型
             TypeCode tc = Type.GetTypeCode(mx.Type);
-            if (tc >= TypeCode.Int16 && tc <= TypeCode.UInt64) return mx;
+            if (tc >= TypeCode.SByte && tc <= TypeCode.UInt64) return mx;
 
             return null;
         }
